Refactored ProgramCardTest and ProgramDeckTest so they are in the correct folder
